requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
python-dotenv==1.0.1
SQLAlchemy==2.0.30
typing_extensions==4.12.2
Werkzeug==3.0.3
